requirements.txt
Django==2.0.9
django-crispy-forms==1.7.2
django-extensions==2.1.3
django-filter
django-ckeditor
djangorestframework==3.9.0
django-rest-auth
pillow
